fix vercel buiding issue by solving File API
